fix: divide components for cookies
